refactor(container): split container.py into a package (#2542, task 1)
Pure move, no behavior changes: the 2792-line flat module becomes
klangk/container/ with one module per collaborator —

- naming.py    _workspace_name_slug/_workspace_container_name
- state.py     ContainerState
- ports.py     PortAllocator + CONTAINER_PORT_START/DEFAULT_PORTS_
               PER_WORKSPACE
- browsers.py  BrowserRouter
- idle.py      IdleMonitor
- health.py    HealthMonitor + unhealthy_message
- sidecar.py   network sidecar lifecycle as NetworkSidecarMixin
               (methods keep their self-registry shape so every
               registry.<sidecar-method> call site is unchanged) +
               the _NETWORK_SIDECAR_* constants
- registry.py  ContainerRegistry (now mixing in the sidecar) with
               bringup, mounts/env, reaps, shutdown
- __init__.py  re-exports the old public surface (incl. ssl_env_vars
               and the private names tests/siblings import), so
               klangk.container.<name> keeps resolving

Test adjustments are path-only: monkeypatch targets that must hit the
defining module's globals now point at klangk.container.registry
(_pid_alive, os.kill) or klangk.container.sidecar
(_NETWORK_SIDECAR_READY_*, os.listdir/unlink). caplog logger names
still work (child loggers inherit the klangk.container level).

4970 passed, coverage 100%.
