+++ src/klangk/klangk/container/__init__.py
@@ -1,0 +1,46 @@
+"""Container lifecycle management package (#2542 split of container.py).
+
+The old single 2792-line ``klangk/container.py`` was split into
+single-responsibility modules; this package ``__init__`` re-exports the
+previous public surface (plus the private names tests and sibling modules
+patch/import) so callers keep working unchanged:
+
+- :mod:`.naming`   - container-name helpers
+- :mod:`.state`    - ``ContainerState``
+- :mod:`.ports`    - ``PortAllocator`` + port constants
+- :mod:`.browsers` - ``BrowserRouter``
+- :mod:`.idle`     - ``IdleMonitor``
+- :mod:`.health`   - ``HealthMonitor`` + ``unhealthy_message``
+- :mod:`.sidecar`  - network sidecar lifecycle (mixin + constants)
+- :mod:`.registry` - ``ContainerRegistry`` (lifecycle, bringup, reaps)
+
+Patch-target note: monkeypatching module globals must now target the
+defining submodule (e.g. ``klangk.container.registry._pid_alive``,
+``klangk.container.sidecar._NETWORK_SIDECAR_READY_TIMEOUT``); re-exports
+here are bindings, not live cells.
+"""
+
+from ..ssl_trust import ssl_env_vars as ssl_env_vars
+from .browsers import BrowserRouter as BrowserRouter
+from .health import (
+    HEALTH_MESSAGE_MAX_BYTES as HEALTH_MESSAGE_MAX_BYTES,
+    HealthMonitor as HealthMonitor,
+    unhealthy_message as unhealthy_message,
+)
+from .idle import IdleMonitor as IdleMonitor
+from .naming import (
+    _workspace_container_name as _workspace_container_name,
+    _workspace_name_slug as _workspace_name_slug,
+)
+from .ports import (
+    CONTAINER_PORT_START as CONTAINER_PORT_START,
+    DEFAULT_PORTS_PER_WORKSPACE as DEFAULT_PORTS_PER_WORKSPACE,
+    PortAllocator as PortAllocator,
+)
+from .registry import (
+    ContainerRegistry as ContainerRegistry,
+    _pid_alive as _pid_alive,
+    _reap_sort_key as _reap_sort_key,
+)
+from .sidecar import NetworkSidecarMixin as NetworkSidecarMixin
+from .state import ContainerState as ContainerState
