+++ src/klangk/klangk/container/registry.py
@@ -1,82 +1,38 @@
-"""Container lifecycle management: start, stop, idle timeout, port allocation."""
+"""Container lifecycle management: start, stop, bringup, reaps (#2542 split).
+
+``ContainerRegistry`` composes :class:`PortAllocator`,
+:class:`BrowserRouter`, :class:`IdleMonitor`, and :class:`HealthMonitor`
+as collaborators, and mixes in :class:`NetworkSidecarMixin` for the FQDN
+egress sidecar.  Constructed once in :func:`build_app` and stored on
+``app.state.container_registry`` (#1426).
+"""
 
 import asyncio
 import logging
 import os
-import re
 import time
 
-from . import podman
-from . import workspace_settings as ws_settings
-from .model.workspaces import EGRESS_MODE_ALLOW, EGRESS_MODE_INTERACTIVE
-from .podman import PodmanError
-
-logger = logging.getLogger(__name__)
-
-
-# --- Runtime SSL/CA certificate injection (#1181) -------------------------
-from .ssl_trust import (  # noqa: E402
+from .. import podman
+from .. import workspace_settings as ws_settings
+from ..model.workspaces import EGRESS_MODE_ALLOW, EGRESS_MODE_INTERACTIVE
+from ..podman import PodmanError
+from ..ssl_trust import (  # noqa: F401 (re-exported via the package)
     SSL_MOUNT_DEST as _SSL_MOUNT_DEST,
     ssl_env_vars,
 )
+from .browsers import BrowserRouter
+from .health import HealthMonitor
+from .idle import IdleMonitor
+from .naming import _workspace_container_name, _workspace_name_slug
+from .ports import (
+    CONTAINER_PORT_START,
+    DEFAULT_PORTS_PER_WORKSPACE,
+    PortAllocator,
+)
+from .sidecar import NetworkSidecarMixin
+from .state import ContainerState
 
-
-# ---------------------------------------------------------------------------
-# Pure module-level constants (no settings reads — #1487 carve-out)
-# ---------------------------------------------------------------------------
-
-CONTAINER_PORT_START = 8000
-DEFAULT_PORTS_PER_WORKSPACE = 5
-
-HEALTH_MESSAGE_MAX_BYTES = 512
-
-# Network sidecar readiness (#2277): the sidecar's proxy prints
-# "dns-proxy listening" once bound. _start_network_sidecar polls the sidecar's
-# logs for it before returning, so a workspace never joins a netns whose OUTPUT
-# policy is still ACCEPT (the entrypoint hasn't run -P OUTPUT DROP yet) — a
-# fail-open window. Module-level so the timeout path is unit-testable fast.
-_NETWORK_SIDECAR_READY_TIMEOUT = 30.0
-_NETWORK_SIDECAR_READY_POLL = 0.3
-_NETWORK_SIDECAR_READY_TOKEN = "dns-proxy listening"
-# fwmark the sidecar's proxy stamps on its upstream socket and the entrypoint
-# matches in its nat/filter rules (#2264). Single source of truth, passed to the
-# sidecar via KLANGKNETWORK_EGRESS_MARK so proxy.py and entrypoint.sh (which
-# both default to 75) can't diverge (#2282).
-_NETWORK_SIDECAR_MARK = 75
-# NFQUEUE queue number the sidecar's interactive-mode consumer binds (#2242).
-# Matches the old NFLOG group for familiarity; single source of truth passed to
-# both the entrypoint (iptables -j NFQUEUE) and the consumer via env.
-_NETWORK_SIDECAR_NFQUEUE = 5139
-# How often the idle cleanup loop scans data_dir/ws-tokens/ for orphaned
-# sidecar-token files (workspaces whose DB row is gone). Self-throttled so
-# the directory + workspace table are not hit on every idle tick (#2309).
-ORPHAN_TOKEN_SWEEP_INTERVAL = 300
-
-
-def _workspace_name_slug(name: str, *, limit: int = 24) -> str:
-    """Sanitize a workspace name for embedding in a container name (#2286).
-
-    Podman container names must match ``[a-zA-Z0-9][a-zA-Z0-9_.-]*``. Lowercase,
-    collapse runs of non-``[a-z0-9]`` to a single ``-``, trim the ends, and cap
-    the length. Returns ``""`` for names that reduce to nothing (empty,
-    all-symbols) — callers fall back to an id-only name. The slug is decorative:
-    uniqueness always comes from the instance id + workspace id, never the slug.
-    """
-    slug = re.sub(r"[^a-z0-9]+", "-", (name or "").lower().strip())
-    return slug[:limit].strip("-")
-
-
-def _workspace_container_name(iid: str, workspace_id: str, slug: str) -> str:
-    """The workspace container name: iid + slugified name + id[:8] (#2286).
-
-    Falls back to an id-only name when the slug is empty (all-symbol / missing
-    name). Uniqueness is iid + id; the slug is decorative. The network sidecar
-    name (:meth:`ContainerRegistry._network_sidecar_name`) shares the id[:8]
-    tail so an id-prefix grep matches the pair.
-    """
-    if slug:
-        return f"klangk-{iid}-{slug}-{workspace_id[:8]}"
-    return f"klangk-{iid}-{workspace_id[:8]}"
+logger = logging.getLogger(__name__)
 
 
 _VALID_PULL_POLICIES = {"never", "missing", "always", "newer"}
@@ -154,545 +110,7 @@ def _reap_sort_key(c: dict) -> int:
     )
 
 
-class ContainerState:
-    """Per-workspace container lifecycle state."""
-
-    def __init__(self, workspace_id: str, container_id: str, app):
-        self.workspace_id = workspace_id
-        self.container_id = container_id
-        self.app = app
-        self.last_activity = time.time()
-        self.idle_timeout: int | None = None
-        self.idle_callbacks: list = []
-        # Health-monitoring state (#1015).  Populated at container start
-        # time so HealthMonitor can poll without a DB lookup per tick.
-        self.health_status: str | None = None  # "healthy" | "unhealthy"
-        self.health_checked_at: float | None = None  # time.time() of last
-        # Short, human-readable reason for the last unhealthy result
-        # (stderr/stdout tail or exception text).  None when healthy or
-        # not yet checked.  Surfaced via the status API + service_health
-        # event so an unhealthy workspace isn't a black box (#1088).
-        self.health_message: str | None = None
-        # Per-workspace monotonic counter carried on every service_health
-        # frame so a reconnecting consumer can detect a missed transition
-        # against the connect-time snapshot (#1175 item 4).  Increments on
-        # each emitted frame (transition and death); resets when the state
-        # is recreated (container restart) -- the snapshot reconciles.
-        self.health_seq: int = 0
-        self.health_check: str | None = None  # shell command, None = disabled
-        self.owner_id: str | None = None
-        self.setup_state: str | None = None
-        # Anchor for the startup grace window
-        # (HEALTH_CHECK_STARTUP_GRACE_SECONDS): the moment the monitored
-        # service began starting.  Defaults to now (container-state
-        # creation) so health-checked workspaces with no service command
-        # still get a grace window; reset to "now" by
-        # ``mark_service_started`` when the service command actually
-        # fires, which is the precise point the service begins booting.
-        self.service_started_at: float = time.time()
-
-    def record_activity(self) -> None:
-        self.last_activity = time.time()
-
-    def mark_service_started(self) -> None:
-        """Record that the service command just fired.
-
-        Resets the startup-grace anchor to now.  Called from
-        ``terminal.ensure_service_session`` right after it launches the
-        service command, so the grace window is measured from the real
-        start of the service rather than from the (earlier) container
-        creation -- important for the per-connection fire path where a
-        freshly-created workspace launches its service command on first
-        ``terminal_start``, possibly long after the container started.
-        """
-        self.service_started_at = time.time()
-
-    def get_idle_timeout(self) -> int:
-        if self.idle_timeout is not None:
-            return self.idle_timeout
-        return self.app.state.container_registry.idle_timeout_seconds
-
-
-class PortAllocator:
-    """Port allocation for workspace containers.
-
-    Owns the ``port_lock`` and delegates to ``model`` for DB-backed
-    port tracking.  Extracted from ``ContainerRegistry`` (issue #972).
-    """
-
-    def __init__(self, app) -> None:
-        self.port_lock: asyncio.Lock = asyncio.Lock()
-        self.app = app
-
-    def reconfigure(self, app) -> None:
-        self.app = app
-
-    async def allocate_ports(self, workspace_id: str, count: int) -> list[int]:
-        # Clamp to the server-wide cap (KLANGKD_HOSTED_PORTS_PER_WORKSPACE)
-        # so creation never allocates ports the deployer has disabled —
-        # otherwise a cap of 0 would still leave orphan allocations
-        # until the container's first start reconcile (#1237).
-        count = min(
-            count, self.app.state.container_registry.ports_per_workspace_cap()
-        )
-        async with self.port_lock:
-            return await self.app.state.model.ports.find_and_allocate_ports(
-                workspace_id,
-                count,
-                self.app.state.container_registry.port_range_start,
-            )
-
-    async def get_workspace_ports(self, workspace_id: str) -> list[int]:
-        return await self.app.state.model.ports.get_workspace_ports(
-            workspace_id
-        )
-
-
-class BrowserRouter:
-    """Browser-delegate routing: browser_id → (workspace_id, sock).
-
-    Browser IDs are browser-generated UUIDs (sessionStorage) sent
-    with terminal_start.  Unlike the old bridge tokens they survive
-    browser refresh because the same sessionStorage UUID re-registers
-    with the new WebSocket.
-
-    Extracted from ``ContainerRegistry`` (issue #972).
-    """
-
-    def __init__(self) -> None:
-        self._browsers: dict[str, tuple[str, object | None]] = {}
-
-    def register_browser(
-        self, browser_id: str, workspace_id: str, sock: object
-    ) -> None:
-        """Register a browser ID for bridge routing.
-
-        Idempotent: the same *browser_id* can re-register with a new
-        *sock* after a browser refresh (sessionStorage keeps the ID).
-        """
-        self._browsers[browser_id] = (workspace_id, sock)
-
-    def resolve_browser(self, browser_id: str) -> tuple[str, object] | None:
-        """Look up (workspace_id, sock) for a browser ID."""
-        return self._browsers.get(browser_id)
-
-    def revoke_workspace_browsers(self, workspace_id: str) -> None:
-        """Remove ALL browser registrations for a workspace.
-
-        Called when a container is recreated or stopped.
-        """
-        to_remove = [
-            bid
-            for bid, (ws, _s) in self._browsers.items()
-            if ws == workspace_id
-        ]
-        for bid in to_remove:
-            del self._browsers[bid]
-
-    def revoke_browser(self, sock: object) -> None:
-        """Remove all browser registrations bound to a specific socket."""
-        to_remove = [
-            bid for bid, (_ws, s) in self._browsers.items() if s is sock
-        ]
-        for bid in to_remove:
-            del self._browsers[bid]
-
-
-class IdleMonitor:
-    """Idle-timeout tracking and cleanup loop.
-
-    Monitors ``ContainerState.last_activity`` for all tracked
-    workspaces and kills containers that exceed their idle timeout.
-
-    Extracted from ``ContainerRegistry`` (issue #972).
-    """
-
-    def __init__(self, app) -> None:
-        self.app = app
-        self.cleanup_task: asyncio.Task | None = None
-        self._cleanup_wake: asyncio.Event | None = None
-
-    def reconfigure(self, app) -> None:
-        self.app = app
-
-    def get_cleanup_wake(self) -> asyncio.Event:
-        if self._cleanup_wake is None:
-            self._cleanup_wake = asyncio.Event()
-        return self._cleanup_wake
-
-    def on_idle_stop(self, workspace_id: str, callback) -> None:
-        state = self.app.state.container_registry.states.get(workspace_id)
-        if state:
-            state.idle_callbacks.append(callback)
-
-    def remove_idle_callback(self, workspace_id: str, callback) -> None:
-        state = self.app.state.container_registry.states.get(workspace_id)
-        if state and callback in state.idle_callbacks:
-            state.idle_callbacks.remove(callback)
-
-    def set_workspace_idle_timeout(
-        self, workspace_id: str, seconds: int
-    ) -> None:
-        state = self.app.state.container_registry.states.get(workspace_id)
-        if state:
-            state.idle_timeout = seconds
-            self.get_cleanup_wake().set()
-
-    def get_workspace_idle_timeout(self, workspace_id: str) -> int:
-        state = self.app.state.container_registry.states.get(workspace_id)
-        if state:
-            return state.get_idle_timeout()
-        return self.app.state.container_registry.idle_timeout_seconds
-
-    async def cleanup_idle_containers(self) -> None:
-        registry = self.app.state.container_registry
-        last_token_sweep = 0.0
-        while True:
-            timeouts = [
-                s.idle_timeout
-                for s in registry.states.values()
-                if s.idle_timeout is not None
-            ]
-            if timeouts:
-                interval = max(2, min(timeouts) // 2)
-            else:
-                interval = registry.check_interval_seconds
-            wake = self.get_cleanup_wake()
-            wake.clear()
-            try:
-                await asyncio.wait_for(wake.wait(), timeout=interval)
-            except asyncio.TimeoutError:
-                pass
-            now = time.time()
-            to_stop = []
-            for ws_id, state in list(registry.states.items()):
-                timeout = state.get_idle_timeout()
-                idle_secs = now - state.last_activity
-                logger.debug(
-                    "Idle check: %s idle %.0fs / %ds",
-                    state.container_id[:12],
-                    idle_secs,
-                    timeout,
-                )
-                if timeout > 0 and idle_secs > timeout:
-                    to_stop.append((state.container_id, ws_id))
-
-            for cid, wid in to_stop:
-                logger.info(
-                    "Stopping idle container %s (workspace %s)",
-                    cid,
-                    wid,
-                )
-                state = registry.states.get(wid)
-                if state:
-                    for cb in list(state.idle_callbacks):
-                        try:
-                            await cb(wid)
-                        except Exception as e:
-                            logger.error("Idle callback error: %s", e)
-                await registry.notify_workspace_killed(wid)
-                await registry.stop_and_remove_container(cid)
-
-            # Periodic orphan sidecar-token sweep (#2309): reclaim
-            # ws-tokens/<id> files whose workspace row is gone. Piggybacks
-            # on this loop (no separate task); self-throttled to scan at
-            # most every ORPHAN_TOKEN_SWEEP_INTERVAL.
-            if now - last_token_sweep >= ORPHAN_TOKEN_SWEEP_INTERVAL:
-                last_token_sweep = now
-                try:
-                    await registry._sweep_orphaned_sidecar_tokens()
-                except Exception as e:
-                    logger.warning("Orphan sidecar-token sweep failed: %s", e)
-
-    def start_cleanup_loop(self) -> None:
-        registry = self.app.state.container_registry
-        logger.info(
-            "Instance: %s, idle timeout: %ds, check interval: %ds",
-            self.app.state.util.instance_id(),
-            registry.idle_timeout_seconds,
-            registry.check_interval_seconds,
-        )
-        if self.cleanup_task is None:
-            self.cleanup_task = asyncio.create_task(
-                self.cleanup_idle_containers()
-            )
-
-
-def unhealthy_message(rc: int, out: str, err: str) -> str:
-    """Build a bounded failure reason from a check's exit code/output.
-
-    Prefers stderr (where shells/diagnostics write their failures);
-    falls back to a tail of stdout; if both are empty, reports just
-    the exit code.  Truncated to ``HEALTH_MESSAGE_MAX_BYTES`` so a
-    verbose check can't grow memory unbounded across workspaces --
-    the goal is "why did it fail", not a full transcript (#1088).
-    """
-    body = (err or "").strip() or (out or "").strip()
-    if body and len(body) > HEALTH_MESSAGE_MAX_BYTES:
-        body = "..." + body[-HEALTH_MESSAGE_MAX_BYTES:]
-    return f"exited {rc}: {body}" if body else f"exited {rc}"
-
-
-class HealthMonitor:
-    """Periodically poll container service health via ``podman exec``.
-
-    Mirrors :class:`IdleMonitor` in shape.  For each workspace with a
-    running container and a configured ``health_check`` command, runs
-    the command inside the container as the creating user with their
-    HOME set.  Exit 0 -> healthy; anything else (non-zero, timeout,
-    error) -> unhealthy.  Status transitions are broadcast to
-    connected clients as ``service_health`` events.
-
-    See #1015 for the design rationale (external polling beats a
-    container-side WS agent here).
-    """
-
-    def __init__(self, app) -> None:
-        self.app = app
-        self.health_task: asyncio.Task | None = None
-
-    def reconfigure(self, app) -> None:
-        self.app = app
-
-    @property
-    def connections(self):
-        """The WebSocketState instance the monitor broadcasts through (#1464).
-
-        Reached via ``app_state.sockets`` (owned instance, #1426) — no
-        post-construction wiring needed.
-        """
-        return self.app.state.sockets
-
-    def _setup_complete(self, state: ContainerState) -> bool:
-        """True if health checks may run for this workspace.
-
-        Checks are skipped until setup has finished (setup_state ==
-        "complete"); running them during setup would report false
-        negatives (the service isn't running yet because setup.sh
-        hasn't installed it).
-        """
-        return state.setup_state == "complete"
-
-    def _in_startup_grace(self, state: ContainerState) -> bool:
-        """True while the service is still within its startup grace window.
-
-        Mirrors Docker's HEALTHCHECK ``--start-period``: while the
-        service command is booting, a failing check is expected rather
-        than a real outage, so :meth:`_check_workspace` ignores
-        unhealthy results here (but still records a *healthy* result so
-        a fast-booting service is marked up immediately).  Anchored to
-        ``service_started_at`` (when the command fired, or container
-        creation as a fallback).
-        """
-        return (
-            time.time() - state.service_started_at
-            < self.app.state.container_registry.health_check_startup_grace
-        )
-
-    async def _run_one(self, state: ContainerState) -> tuple[str, str]:
-        """Run a single workspace's health check.
-
-        Returns ``(status, message)`` where *status* is ``"healthy"`` or
-        ``"unhealthy"`` and *message* is a short, human-readable reason
-        for an unhealthy result (a bounded tail of the check's
-        stderr/stdout, or the exception text) -- empty when healthy.
-        Surfacing the reason turns an ``unhealthy`` status from a black
-        box into a diagnosable failure instead of "good luck" (#1088).
-
-        Resolves the owner's container home (same logic as
-        ``start_workspace``) and invokes the check via
-        ``podman exec`` as the creating user with HOME set.  The check
-        runs as a **non-login** bash shell (``bash -c``) on purpose: it
-        is an operational probe, not a user session, so it deliberately
-        sources *no* startup file -- not ``~/.profile``, ``~/.bashrc``,
-        nor ``/etc/profile.d/*``.  This keeps the probe deterministic
-        and decoupled from the owning user's interactive setup: a slow
-        ``nvm`` load, a broken ``~/.profile`` edit, or a stray ``read``
-        prompt must never make an unattended 30s poll flap "unhealthy".
-
-        The flip side is that the check command must not rely on the
-        user's PATH or env.  It inherits only the container's image
-        ``PATH`` (so ``/opt/klangk/bin`` and system tools like
-        ``grep``/``curl`` resolve) plus ``HOME``.  Anything the checked
-        service needs -- a sandbox-installed binary, ``OPENCLAW_HOME`` /
-        ``HERMES_HOME``, a custom ``PATH`` -- must be referenced by
-        **absolute path** in the check command, or wrapped in an
-        executable script whose shebang and ``export`` lines bake those
-        in (the recommended pattern for non-trivial checks; see
-        ``docs/features/health-check.md``).  Errors and timeouts count
-        as ``"unhealthy"``.
-        """
-        owner_id = state.owner_id
-        if owner_id is None:
-            return "unhealthy", "no owner recorded for workspace"
-        handle = await self.app.state.model.users.get_user_handle(owner_id)
-        if not handle:
-            return "unhealthy", f"owner {owner_id} has no handle"
-        # Resolve the owner's container home the same way
-        # start_workspace does, so the check runs in the right
-        # HOME rather than as root in /.
-        ws = self.app.state.workspaces
-        ws_home = ws.home_path(state.workspace_id)
-        user_home, _created = await ws.ensure_home_symlink(
-            ws_home, handle, owner_id
-        )
-        cid_short = state.container_id[:12]
-        logger.debug(
-            "Health check: container %s (workspace %s) running %r",
-            cid_short,
-            state.workspace_id,
-            state.health_check,
-        )
-        try:
-            rc, out, err = await self.app.state.podman.exec_container(
-                state.container_id,
-                # bash -c (NON-login): sources nothing, so the probe is
-                # deterministic and insulated from the user's interactive
-                # shell setup. Only the image PATH + HOME are visible,
-                # so health_check commands must use absolute paths (or a
-                # wrapper script). See docs/features/health-check.md.
-                # Skipped until setup_state == complete.
-                ["bash", "-c", state.health_check],
-                user="klangk",
-                extra_env={"HOME": user_home},
-                timeout=self.app.state.container_registry.health_check_timeout,
-            )
-        except (podman.PodmanError, asyncio.TimeoutError, OSError) as e:
-            return "unhealthy", f"{type(e).__name__}: {e}"
-        if rc == 0:
-            return "healthy", ""
-        return "unhealthy", unhealthy_message(rc, out, err)
-
-    async def _check_workspace(self, state: ContainerState) -> None:
-        """Poll one workspace, record the reason, and broadcast on change."""
-        new_status, message = await self._run_one(state)
-        # Startup grace window: the service command may still be
-        # booting, so an unhealthy result here is expected, not a real
-        # outage.  Don't transition to unhealthy, broadcast, or log a
-        # failure (mirrors Docker HEALTHCHECK --start-period).  A
-        # healthy result is still recorded below so a fast-booting
-        # service is marked up the moment it actually responds.
-        if new_status == "unhealthy" and self._in_startup_grace(state):
-            logger.debug(
-                "Health check for workspace %s (container %s) failing "
-                "but within startup grace (%.0fs elapsed); not flagging "
-                "unhealthy",
-                state.workspace_id,
-                state.container_id[:12],
-                time.time() - state.service_started_at,
-            )
-            return
-        old_status = state.health_status
-        state.health_status = new_status
-        # Clear the reason once healthy again so a stale failure message
-        # can't linger next to a "healthy" status (#1088).
-        state.health_message = message if new_status == "unhealthy" else None
-        state.health_checked_at = time.time()
-        if new_status == "unhealthy":
-            # Log the reason at info on a fresh transition (so it's
-            # visible without debug logs), debug on steady-state polls
-            # so a persistently-broken check doesn't spam at info (#1088).
-            log = logger.info if old_status != "unhealthy" else logger.debug
-            log(
-                "Health check for workspace %s (container %s) unhealthy: %s",
-                state.workspace_id,
-                state.container_id[:12],
-                message,
-            )
-        if new_status != old_status:
-            self._broadcast(state, new_status, state.health_message)
-
-    def _broadcast(
-        self,
-        state: ContainerState,
-        status: str,
-        message: str | None = None,
-    ) -> None:
-        """Emit a ``service_health`` transition event to all connections.
-
-        Fanned out via :meth:`WsState.notify_service_health` so the
-        workspace list page learns about health transitions for
-        auto-started services even when nobody is connected to the
-        workspace's terminal session (#1015).  The failure *reason*
-        rides along as ``health_message`` so operators can see *why*
-        it's unhealthy without digging through logs (#1088).
-
-        Also forwards the additive contract fields (#1175):
-        ``running=True`` (this is a live-container frame), the last
-        ``health_checked_at`` (#1175 item 3a), and a per-workspace
-        ``seq`` (#1175 item 4) bumped on every emit so a reconnecting
-        consumer can detect a missed transition.
-        """
-        state.health_seq += 1
-        self.connections.notify_service_health(
-            state.workspace_id,
-            healthy=status == "healthy",
-            message=message,
-            running=True,
-            health_checked_at=state.health_checked_at,
-            seq=state.health_seq,
-        )
-
-    def broadcast_death(self, state: ContainerState) -> None:
-        """Emit the terminal ``service_health`` frame for a dying container.
-
-        When a container dies the server emits
-        ``container_status{running: false}`` and then *silence* on the
-        ``service_health`` stream, because the health loop only polls
-        ``registry.states`` and a dead container's state is removed.  A
-        consumer watching ``service_health`` therefore believes the
-        last-known status (possibly healthy) still holds while the
-        container is gone (#1175 item 2).  This closes the hole by
-        emitting one unambiguous terminal frame with ``running=False``
-        and ``healthy=False`` *before* the state is dropped, so a single
-        stream is a single source of truth.
-        """
-        state.health_seq += 1
-        self.connections.notify_service_health(
-            state.workspace_id,
-            healthy=False,
-            message=None,
-            running=False,
-            health_checked_at=state.health_checked_at,
-            seq=state.health_seq,
-        )
-
-    async def run_health_loop(self) -> None:
-        """Background loop: every interval, poll eligible workspaces.
-
-        After each poll sweep, emits liveness heartbeats to connections
-        that opted in (#1175 item 3b).  Emitting from *this* loop (rather
-        than a standalone task) ties heartbeat presence to the health
-        loop being alive -- if the loop stalls, the heartbeats stop.
-        """
-        while True:
-            registry = self.app.state.container_registry
-            await asyncio.sleep(registry.health_check_interval)
-            for state in list(registry.states.values()):
-                if not state.health_check:
-                    continue
-                if not self._setup_complete(state):
-                    continue
-                try:
-                    await self._check_workspace(state)
-                except Exception as e:  # pragma: no cover - defensive
-                    logger.error(
-                        "Health check error for workspace %s: %s",
-                        state.workspace_id,
-                        e,
-                    )
-            self._send_heartbeats()
-
-    def _send_heartbeats(self) -> None:
-        """Fan health heartbeats to opt-in connections."""
-        self.connections.send_health_heartbeats()
-
-    def start_health_loop(self) -> None:
-        if self.health_task is None:
-            self.health_task = asyncio.create_task(self.run_health_loop())
-
-
-class ContainerRegistry:
+class ContainerRegistry(NetworkSidecarMixin):
     """Manages all container state and podman interactions.
 
     Composes :class:`PortAllocator`, :class:`BrowserRouter`,
@@ -1207,354 +625,6 @@ class ContainerRegistry:
             {"settings": workspace_settings},
             self.app.state.settings.container_tmp_size,
         )
-
-    def _network_sidecar_enabled(self) -> bool:
-        """Whether the FQDN network sidecar model is configured (#2254).
-
-        Mirrors :meth:`NetFilter.enabled`: the master switch on AND the
-        sidecar image set (which ships with a default, so this is True out
-        of the box)."""
-        return self.app.state.settings.netfilter_enabled and bool(
-            self.app.state.settings.network_sidecar_image
-        )
-
-    def _network_sidecar_name(self, workspace_id: str, slug: str = "") -> str:
-        """Derive the network sidecar name (#2254, #2286).
-
-        Carries the slugified workspace name (when there is one) so
-        ``podman ps | grep <partial-name>`` surfaces the sidecar next to its
-        workspace, and uses the same ``workspace_id[:8]`` tail as the workspace
-        container name so an id-prefix grep matches the pair. Removal is by the
-        ``klangk.workspace`` label (see :meth:`_remove_network_sidecar`), not by
-        name — so a stale slug (after a rename) can't strand an orphan.
-        """
-        if slug:
-            return f"klangk-net-{slug}-{workspace_id[:8]}"
-        return f"klangk-net-{workspace_id[:8]}"
-
-    def _sidecar_token_path(self, workspace_id: str) -> str:
-        """Host path of the sidecar's bind-mounted workspace-token file."""
-        return os.path.join(
-            self.app.state.settings.data_dir, "ws-tokens", workspace_id
-        )
-
-    def write_sidecar_token(self, workspace_id: str, token: str) -> None:
-        """Write the current workspace token to the sidecar's token file.
-
-        The sidecar reads this (read-only bind mount) on each consent POST, so
-        refreshing it here -- on launch and on every WS-driven token rotation
-        -- keeps the sidecar authenticated without baking a (rotating,
-        expiring) token into its env (#2242). The sidecar image lacks the
-        workspace image's token-setter, so it can't be exec-pushed like the
-        workspace container. Atomic (os.replace) so the sidecar never reads a
-        half-written token.
-        """
-        path = self._sidecar_token_path(workspace_id)
-        os.makedirs(os.path.dirname(path), exist_ok=True)
-        tmp = f"{path}.tmp"
-        with open(tmp, "w") as f:
-            f.write(token)
-        os.replace(tmp, path)
-
-    async def _sweep_orphaned_sidecar_tokens(self) -> int:
-        """Remove ``ws-tokens/<id>`` files for workspaces that no longer exist.
-
-        The sidecar's workspace-token file is written on every network-sidecar
-        start (:meth:`write_sidecar_token`) and intentionally kept across
-        stop/restart -- a stopped workspace must keep its token for a later
-        restart. Only a final delete (or a crash that skips cleanup) orphans
-        the file, so this sweep reclaims them (#2309). Safe: a workspace whose
-        row still exists keeps its token; only files with no matching row are
-        removed. Returns the number of files removed.
-        """
-        token_dir = os.path.join(self.app.state.settings.data_dir, "ws-tokens")
-        if not os.path.isdir(token_dir):
-            return 0
-        try:
-            names = os.listdir(token_dir)
-        except OSError as e:
-            logger.warning(
-                "Cannot scan %s for orphaned tokens: %s", token_dir, e
-            )
-            return 0
-        try:
-            existing = await self.app.state.workspaces.existing_workspace_ids()
-        except Exception as e:
-            logger.warning("Cannot list workspace ids for token sweep: %s", e)
-            return 0
-        removed = 0
-        for name in names:
-            # Skip transient write_sidecar_token temp files: unlinking one
-            # races the writer (os.replace would then FileNotFoundError).
-            if name.endswith(".tmp"):
-                continue
-            path = os.path.join(token_dir, name)
-            if not os.path.isfile(path):
-                continue
-            if name in existing:
-                continue
-            try:
-                os.unlink(path)
-                removed += 1
-                logger.info("Removed orphaned sidecar token %s", name)
-            except OSError as e:
-                logger.warning(
-                    "Could not remove orphaned token %s: %s", name, e
-                )
-        return removed
-
-    async def _start_network_sidecar(
-        self,
-        workspace_id: str,
-        allowed_domains: list[str] | None,
-        egress_mode: str = "static",
-        rejected_domains: list[str] | None = None,
-        publish: list[tuple[int, int]] | None = None,
-        slug: str = "",
-    ) -> str:
-        """Create + start the FQDN network sidecar for a filtered workspace (#2254).
-
-        Returns the network sidecar's container ID (for ``--network container:<id>``).
-        Raises ``podman.PodmanError`` on failure — the caller fail-closes (a
-        workspace that declared an allow-list never starts unrestricted; #2254
-        review B2). The network sidecar gets ``--cap-add NET_ADMIN`` + ``--dns 1.1.1.1``
-        (the REDIRECT target the workspace inherits); the proxy forwards to a
-        *different* detected upstream (loop avoidance). The allow-list + the
-        klangkd backend port are passed via env. ``publish`` (the host ports the
-        workspace requested) is published on the sidecar itself (#2267): the
-        workspace shares the sidecar's netns, so ``--publish`` is inert on the
-        workspace under ``--network container:`` but works on the netns owner,
-        forwarding into the shared netns to the workspace's listener — letting
-        filtered workspaces host apps.
-        """
-        image = self.app.state.settings.network_sidecar_image
-        if not image:
-            raise podman.PodmanError(
-                500, "network_sidecar_image is not configured"
-            )
-        name = self._network_sidecar_name(workspace_id, slug)
-        # Pick an upstream that differs from the REDIRECT target (1.1.1.1).
-        # KLANGKNETWORK_EGRESS_UPSTREAM (when set in klangkd's env) pins the
-        # sidecar's upstream resolver verbatim, mirroring the MIN_TTL /
-        # SWEEP_INTERVAL forwarding below — an operator may want workspaces to
-        # use a specific resolver (e.g. a corporate DNS), and the egress
-        # smoketest uses it to point the sidecar at a controlled-DNS fixture
-        # (#2424) so chosen hostnames resolve to single stable test IPs.
-        # Absent -> auto-detect a host resolver that differs from 1.1.1.1.
-        env_upstream = os.environ.get("KLANGKNETWORK_EGRESS_UPSTREAM")
-        if env_upstream:
-            upstream = env_upstream
-        else:
-            nf = getattr(self.app.state, "netfilter", None)
-            resolvers = nf.resolvers() if nf else []
-            upstream = next(
-                (r for r in resolvers if r != "1.1.1.1"), "8.8.8.8"
-            )
-        env = [
-            f"KLANGKNETWORK_EGRESS_ALLOW={','.join(allowed_domains or [])}",
-            f"KLANGKNETWORK_EGRESS_REJECT={','.join(rejected_domains or [])}",
-            f"KLANGKNETWORK_EGRESS_UPSTREAM={upstream}",
-            # The klangkd backend port (LLM proxy + bridge on
-            # host.containers.internal). The network sidecar allow-lists it statically
-            # — it's a /etc/hosts entry the FQDN proxy can't learn (#2254 B1).
-            f"KLANGKNETWORK_EGRESS_BACKEND_PORT={self.app.state.settings.egress_port}",
-            # Single source of truth for the fwmark both proxy.py and
-            # entrypoint.sh use (#2264, #2282): they default to 75, but pass it
-            # explicitly so the two can't diverge.
-            f"KLANGKNETWORK_EGRESS_MARK={_NETWORK_SIDECAR_MARK}",
-        ]
-        # Forward the learned-IP TTL floor + sweep cadence when the operator
-        # (or a test) sets them. The sidecar defaults (MIN_TTL=30s, sweep=5s,
-        # proxy.py) floor any verdict TTL, so the egress smoketest lowers both
-        # to make a short timed verdict expire in seconds. Absent -> defaults.
-        # ACTIVITY_GATE (default 60s) likewise: the idle fuzz harness
-        # (scripts/fuzz-idle.py, #2514) lowers it so the sidecar's
-        # egress-activity bumps are observable at its seconds-scale timeouts.
-        for _opt in (
-            "KLANGKNETWORK_EGRESS_MIN_TTL",
-            "KLANGKNETWORK_EGRESS_SWEEP_INTERVAL",
-            "KLANGKNETWORK_EGRESS_DEBUG_RST",
-            "KLANGKNETWORK_EGRESS_ACTIVITY_GATE",
-        ):
-            _v = os.environ.get(_opt)
-            if _v:
-                env.append(f"{_opt}={_v}")
-        binds = []
-        # #2242: consent recording (deny + record) runs for every filtered
-        # workspace when the monitor is wired, regardless of egress_mode -- the
-        # mode only affects the recorded decision (static=denied+no-human,
-        # interactive=pending), applied by the monitor. The workspace JWT is
-        # bind-mounted in and refreshed on rotation (write_sidecar_token), not
-        # baked in env (it rotates).
-        if getattr(self.app.state, "consent_monitor", None) is not None:
-            port = self.app.state.settings.egress_port
-            env.append(
-                "KLANGKNETWORK_EGRESS_CONSENT_URL="
-                f"http://host.containers.internal:{port}"
-                "/internal/egress-consent/events"
-            )
-            env.append(
-                f"KLANGKNETWORK_EGRESS_NFQUEUE_NUM={_NETWORK_SIDECAR_NFQUEUE}"
-            )
-            token = self.app.state.auth.create_workspace_token(workspace_id)
-            self.write_sidecar_token(workspace_id, token)
-            binds.append(
-                f"{self._sidecar_token_path(workspace_id)}"
-                ":/run/klangk/workspace-token:ro"
-            )
-        # Label the network sidecar with this klangk instance so the startup reaper
-        # (reap_instance_containers) and the shutdown orphan sweep cull any
-        # network sidecar left behind by a failed stop — the same culling workspace
-        # containers get (#2254 review). #2342: klangk.managed + klangk.pid also
-        # let the dead-owner reap (reap_dead_owner_containers) cull a sidecar
-        # whose creating klangkd died uncleanly, the same as a workspace container.
-        labels = {
-            "klangk.managed": "true",
-            "klangk.instance": self.app.state.util.instance_id(),
-            # The main klangkd daemon process's PID — the liveness signal the
-            # dead-owner reap keys on (#2342). Not conmon / the container's
-            # PID 1 / a podman subprocess: those die with the daemon anyway.
-            "klangk.pid": str(os.getpid()),
-            # #2286: a shared klangk.workspace label + a klangk.role label let
-            # one `podman ps --filter label=klangk.workspace=<id>` correlate
-            # the sidecar with its workspace; the slug is mirrored for
-            # exact-match filtering. (Supersedes the old write-only
-            # klangk.network-sidecar.)
-            "klangk.workspace": workspace_id,
-            "klangk.role": "network-sidecar",
-            "klangk.workspace-name": slug,
-        }
-        # #2265: a sidecar from a prior generation may linger (an unclean
-        # stop, or the workspace container was killed externally so
-        # stop_and_remove_container never ran for it). Clear it before create
-        # so create_container doesn't collide and the caller doesn't
-        # fail-closed. Removal is by the klangk.workspace label (#2286), not
-        # the name — the name now carries the workspace-name slug, which may
-        # differ from a prior generation's (a rename) and can't be
-        # reconstructed from the id alone. The instance reaper is the
-        # backstop for orphans; this just keeps a restart from deadlocking.
-        await self._remove_network_sidecar(workspace_id)
-        try:
-            # NET_RAW lets the proxy forge the eager-deny RST directly from
-            # the NFQUEUE callback (#2345) so a denied connect() gets
-            # ECONNREFUSED at once, independent of the conntrack/retransmit
-            # race that made the iptables REJECT rule flaky. Safe to grant
-            # here (unlike on the workspace): the sidecar runs only klangk's
-            # own proxy, no untrusted code.
-            sidecar_kwargs = dict(
-                cap_add=["NET_ADMIN", "NET_RAW"],
-                dns=["1.1.1.1"],
-                env=env,
-                labels=labels,
-                publish=publish,
-                pull="missing",
-            )
-            if binds:
-                sidecar_kwargs["binds"] = binds
-            cid = await self.app.state.podman.create_container(
-                name, image, **sidecar_kwargs
-            )
-            await self._start_with_port_conflict_retry(
-                cid, publish or [], name
-            )
-            # #2277: wait for the proxy to be listening before returning, so the
-            # workspace never joins a netns whose OUTPUT is still ACCEPT
-            # (entrypoint mid-flight) — a fail-open window. The proxy prints
-            # "dns-proxy listening" once bound; poll its logs. Fail-closed: if
-            # the sidecar exits first or the proxy never binds, raise so the
-            # caller refuses to start the workspace rather than run it unfiltered.
-            deadline = time.monotonic() + _NETWORK_SIDECAR_READY_TIMEOUT
-            ready = False
-            while time.monotonic() < deadline:
-                logs = await self.app.state.podman.container_logs(cid)
-                if _NETWORK_SIDECAR_READY_TOKEN in logs:
-                    ready = True
-                    break
-                state = await self.app.state.podman.inspect_container(cid)
-                status = (state or {}).get("State", {}).get("Status", "")
-                if status in ("exited", "stopped"):
-                    raise podman.PodmanError(
-                        500,
-                        f"network sidecar {name} exited before the DNS proxy "
-                        f"was ready; logs:\n{logs}",
-                    )
-                await asyncio.sleep(_NETWORK_SIDECAR_READY_POLL)
-            if not ready:
-                raise podman.PodmanError(
-                    500,
-                    f"network sidecar {name} DNS proxy did not become ready "
-                    f"within {_NETWORK_SIDECAR_READY_TIMEOUT:.0f}s; the "
-                    "workspace would join an unfiltered netns",
-                )
-            logger.info(
-                "network sidecar started for %s: %s (%s)",
-                workspace_id[:8],
-                name,
-                cid[:12],
-            )
-            return cid
-        except podman.PodmanError as exc:
-            logger.warning(
-                "network sidecar failed for %s: %s — fail-closed at caller",
-                workspace_id[:8],
-                exc,
-            )
-            raise
-
-    async def _stop_network_sidecar(self, workspace_id: str) -> None:
-        """Best-effort remove the network sidecar for a workspace (#2254, #2286).
-
-        Delegates to :meth:`_remove_network_sidecar` (label-based) so a sidecar
-        whose name carries a now-stale slug (a renamed or deleted workspace) is
-        still found and removed.
-        """
-        if not self._network_sidecar_enabled():
-            return
-        await self._remove_network_sidecar(workspace_id)
-
-    async def _remove_network_sidecar(self, workspace_id: str) -> None:
-        """Best-effort remove this workspace's network sidecar by label (#2286).
-
-        Keyed on ``klangk.workspace=<workspace_id>`` +
-        ``klangk.role=network-sidecar`` rather than the container name: the name
-        now carries the slugified workspace name, which can't be reconstructed
-        from the id alone after a rename, a delete (cascade), or a process
-        restart (the in-memory set is gone). Label-based removal finds the
-        sidecar regardless, leaves the workspace container (role=workspace)
-        untouched, and is 404-tolerant.
-        """
-        try:
-            containers = await self.app.state.podman.list_containers(
-                f"klangk.workspace={workspace_id}"
-            )
-        except (podman.PodmanError, OSError, ValueError) as exc:
-            # ValueError: corrupted ps JSON (json.loads in list_containers) —
-            # best-effort removal must never raise into the caller.
-            logger.warning(
-                "cannot list containers to remove network sidecar for %s: %s",
-                workspace_id[:8],
-                exc,
-            )
-            return
-        for c in containers:
-            labels = c.get("Labels") or {}
-            if labels.get("klangk.role") != "network-sidecar":
-                continue
-            ident = c.get("Id") or c.get("ID") or ""
-            if not ident:
-                names = c.get("Names") or []
-                ident = names[0] if names else ""
-            if not ident:
-                continue
-            try:
-                await self.app.state.podman.remove_container(ident, force=True)
-                logger.info(
-                    "network sidecar removed for %s: %s",
-                    workspace_id[:8],
-                    ident[:12],
-                )
-            except podman.PodmanError:
-                pass  # already gone — expected
 
     async def start_container(
         self,
